fix: Fixed data handling in the NaCl client
There are two things to note when receiving data
from the browser. First, the data image is in RGB
format with the alpha channel, so it is RGBA. This
means that the the data type has to be CV_8UC4.

Then, color segmentation and other functions meant
to process the images have to be updated to handle
this new data format.